Complete favorites tab

+++ MovieSearch/MovieSearch.h
@@ -15,7 +15,7 @@
 
 - (void)search:(NSString*)query completion:(void (^)(NSMutableArray*))completion;
 - (NSString *)formatDate:(NSString*)stringDate;
-- (void)getMovieForID:(int)idNumber completion:(void (^)(Movie *))completion;
+- (void)getMovieForID:(NSInteger)idNumber completion:(void (^)(Movie *))completion;
 - (void)getTrailerForID:(NSNumber *)idNumber completion:(void (^)(NSString *))completion;
 - (void)getNowPlaying:(void (^)(NSMutableArray *))completion;
